String commands work with extract_args
#69
