+++ test/test_extract_args.py
@@ -1,10 +1,9 @@
-from fakeredis._commands import extract_args
+from fakeredis._command_args_parsing import extract_args
 
 
 def test_extract_args():
-    xx, nx, ex, keepttl = extract_args(
-        (b'nx', b'ex', b'324', b'xx',),
-        ('nx', 'xx', '+ex', 'keepttl'))
+    args = (b'nx', b'ex', b'324', b'xx',)
+    xx, nx, ex, keepttl = extract_args(args, ('nx', 'xx', '+ex', 'keepttl'))
     assert xx
     assert nx
     assert ex == 324
